seperation of plots, cache dataset
