+++ p2p-frontend/src/components/StaticGraphs.jsx
@@ -8,61 +8,61 @@ function StaticGraphs() {
       <Grid container spacing={4}>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/average-time-by-year'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/average-time-by-year'}
             title="Average Time"
           />
         </Grid>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/median-time-by-year'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/median-time-by-year'}
             title="Median Time"
           />
         </Grid>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/participation-by-year'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/participation-by-year'}
             title="Overall Participation"
           />
         </Grid>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/average-time-by-age-group-and-year'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/average-time-by-age-group-and-year'}
             title="Average Time by Age Group"
           />
         </Grid>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/participation-by-region'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/participation-by-region'}
             title="Regional Participation"
           />
         </Grid>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/time-percentiles-by-year'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/time-percentiles-by-year'}
             title="Time Percentiles"
           />
         </Grid>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/gender-distribution-by-year'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/gender-distribution-by-year'}
             title="Gender Distribution"
           />
         </Grid>
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/top-10-fastest-swimmers'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/top-10-fastest-swimmers'}
             title="Top 10 Fastest Swimmers"
           />
         </Grid>
         {/* <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/age-vs-time-distribution'}
+            apiUrl={'process.env.REACT_APP_API_URI' + '/api/age-vs-time-distribution'}
             title="Age vs Time"
           />
         </Grid> */}
         <Grid item xs={12} sm={6} md={4}>
           <GraphContainer
-            apiUrl={process.env.REACT_APP_API_URI + '/api/participation-by-age-group'}
+            apiUrl={'http://127.0.0.1:5000' + '/api/participation-by-age-group'}
             title="Age Group Participation"
           />
         </Grid>
